+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog of threedigrid-builder
 1.0.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Make requesting spatial reference of GDAL dataset compatible with GDAL 2.x.
 
 
 1.0.2 (2022-01-17)